Fix test coverage of common/Select
